requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-Bcrypt==1.0.1
pymongo[srv]==4.6.1
python-dotenv==1.0.0
openai-whisper